fix order in life-time loop
- error fixed: time updated before animals react in iterate_pop
-first adjustment for free
-in variable runs: no reading of mean and std genes anymore
-lower boundary above 0 for mutation of mu
-choosing of primed or unprimed genes as own method
-std and lineage stop removed

+++ main_variable.py
@@ -54,7 +54,7 @@ from iterate_population import iterate_population
 if __name__ == '__main__':
     # Get model constants
     constants = model_constants
-    population_size=constants["environment_sizes"]
+    population_size=constants["size"]
     
     if constants["path"] !="":
         path=constants["path"]
@@ -63,72 +63,42 @@ if __name__ == '__main__':
         p=open("./path.txt","r")
         path=p.readline()
 
-    f_mean = path + "pop{0}_mean_genes.csv".format(constants["use_pop"]) #specify path to csv file in constants (from run with constant pop size) 
-    f_std = path+ "pop{0}_std_genes.csv".format(constants["use_pop"])  
-    final_state=path+ "pop{0}_final_state.csv".format(constants["use_pop"])  
-
+    final_state=path+ "pop{0}_final_state.csv".format(constants["use_pop"])  #specify path to csv file in constants (from run with constant pop size) 
+    parameters=path+"parameters.txt"
     # create output directory
     now = datetime.datetime.today()
-
-
     start = time.clock()
 
     # read the csv files
-    try:
-        approx=True
-        with open(f_mean) as f:
-            reader = csv.reader(f,delimiter=",")
-            for (i,row) in enumerate(reader):
-                if row and row[0]!="":
-                    if row[0]=="n":
-                        data = np.genfromtxt(f_mean,skip_header=i+1,delimiter=",") #reads mean genes and n , nperPos from csv 
-                        break
-                    elif (row[0][0]!="R"): 
-                        e=row[:5]
-                        environment=list(map(float,e))  #reads environment values 
-                        
-        if constants["trans"]:
-            factor=constants["environments"][0]/environment[0] #to ensure that environment value E is continous (no jump from constant run)
-        else:
-            factor=1
-        final_t = data[-1,0]*constants["L"]*factor
-        size=int(data[-1,-1])
-        data2 = np.genfromtxt(final_state,skip_header=1,delimiter=",") #reads genes and n , nperPos from csv file of the whole final population if available
-        genes1=data2[:,:-1]
-        mean_genes=np.mean(genes1,axis=0)
-        approx=False                
-    except:    #if file for whole population does not exist, use mean values
-        std = np.genfromtxt(f_std,skip_header=i+1,delimiter=",")
-        if len(std[0])==10:
-            std=np.delete(std,3,1) #delete mismatch
-        if len(data[0])==11:
-            data=np.delete(data,(3,10),1) #delete mismatch,lineage
-           # std=np.delete(std,(3,10),1) #delete mismatch,lineage
-        elif len(data[0])==12: #for files with lineage as last entry    
-            data=np.delete(data,(1,8,9),1) #delete environment,m ,ma
-            std=np.delete(std,(1,8,9),1) #delete environment,m ,ma
-        elif len(data[0])==13:
-            data=np.delete(data,(1,4,9,10),1) #delete environment,mismatch, m,ma
-            std=np.delete(std,(1,4,9,10),1) #delete environment,mismatch, m,ma
-        else:
-            raise Exception("Check the delete mechanisms for initial data!")
-        
-        mean_genes = data[-1,1:-1]  #just last row, that is last generation. All genes, cut off generation and size 
-        std_genes = std[-1,1:-1] #std of genes from last generation
-        std_genes = np.fabs(std_genes) #gives absolute value  
-    # create environments and output information about them
+    with open(parameters) as f:
+        for i in range(8):
+            line=f.readline()
+            if i==0:
+                gen=float(line[13:-1]) #number of generations in the constant run
+                
+        line=line[15:-2]
+        env=line.rsplit(",")
+        environment=list(map(float,env))  #reads environment values                 
+                
+    if constants["trans"]:
+        factor=constants["environment"][0]/environment[0] #to ensure that environment value E is continous (no jump from constant run)
+    else:
+        factor=1
+    final_t = gen*constants["L"]*factor #final time of constant run
+    data = np.genfromtxt(final_state,skip_header=1,delimiter=",") #reads genes and n , nperPos from csv file of the whole final population 
+    genes1=data[:,:-1] #last column mismatch is removed
+    mean_genes=np.mean(genes1,axis=0)
+    std_genes=np.std(genes1,axis=0)             
+
+    # create environment and output information about them
 
     if constants["trans"]: #for transition runs, change environment parameters in constants file
-        env = Environment(*constants["environments"]) 
+        env = Environment(*constants["environment"]) 
     else: #or use environment from constant run
         env = Environment(*environment)
         
     if constants["folder"]!="":       
         path = "./output_variable/{0}/".format(constants["folder"])
-        try:
-            os.makedirs(path)      
-        except:
-            pass
     else:
         path = "./output_variable/"
         
@@ -138,7 +108,6 @@ if __name__ == '__main__':
         path=path+constants["desc"]+"/"
     
     try: 
-        os.makedirs(path)
         os.makedirs(path+"timeseries/")
     except OSError:
         if not os.path.isdir(path):
@@ -146,17 +115,11 @@ if __name__ == '__main__':
     f3 = open(path+"__overview.txt",'w')
     f3.write("initial conditions \n")
     f3.write("R,P,A,B,O\n{0},{1},{2},{3},{5}\n".format(env.R,env.P,env.A,env.B,i,env.O))
-    if approx:
-        f3.write("Mean genes(I0,I0p,a,b,bp,h,s):\n{0}\n".format(mean_genes))
-        f3.write("Std genes:\n{0}\n".format(std_genes))
-    else:
-        f3.write("Mean genes(h,s,a,I0,I0p,b,bp,mu):\n{0}\n".format(mean_genes))
-
-    for key in ['generations','L','kd','ka','tau','q','mutation','environments','environment_names','environment_sizes','populations','plot_every','verbose',\
-'random_choice','std_min','lineage_stop','desc','force_plast',"hgt","proc",'trans','path','use_pop']:
+    f3.write("Mean genes(h,s,a,I0,I0p,b,bp,mu):\n{0}\n".format(mean_genes))
+    f3.write("Std genes:\n{0}\n".format(std_genes))
+    for key in ['generations','L','kd','ka','tau','q','mutation','environment','environment_name','size','populations','plot_every','verbose',\
+'random_choice','desc','force_plast',"hgt",'check','kh','kt',"proc",'trans','path','use_pop']:
         f3.write("{0}:\t{1}\n".format(key,constants[key]))
-    
-
     
     end = time.clock()
     if constants["verbose"]:
@@ -179,20 +142,9 @@ if __name__ == '__main__':
             
         # create animals with the mean genes that shall be tested for each environment
         animals=[]
-        
-        # unfortunately, the genes are written in a different order as it is used here
-        if approx:
-            genes = []
-            gene_order = [5,6,2,0,1,3,4] #b,bp,a,h,s,I0,I0p
-            for j in gene_order:
-                if (std_genes[j] > 0):
-                    genes.append(np.random.normal(size=size,loc=mean_genes[j],scale=std_genes[j])) #if std>0 create size*genes using normal distribution around mean for each environment and gene
-                else:
-                    genes.append(mean_genes[j]*np.ones(size))
-            animals.append([Animal(np.array(g),lineage=k) for k,g in enumerate(zip(*genes))])# create animals with genes in environment i
-        else:
-            animals.append([Animal(np.array(g),lineage=k) for k,g in enumerate(genes1)])
+        animals.append([Animal(np.array(g),lineage=k) for k,g in enumerate(genes1[:constants["size"]])])
         animals = [item for sublist in animals for item in sublist] # flatten animal list 
+        
         # create a population of population_size animals that have the correct mean genes
         population = Population(population_size,animals)
         
@@ -201,14 +153,15 @@ if __name__ == '__main__':
 
 
         if population.size()==0:
-            print("\t Population {0} died out! Time needed: {1:.2f} min".format(k,(end-start)/60)) 
+            print("\t Population {0} died out! Time needed: {1:.2f} min".format(k+1,(end-start)/60)) 
             return False,final_gen
         else:
-            print("\t Population {0} survived! Time needed: {1:.2f} min".format(k,(end-start)/60))
+            print("\t Population {0} survived! Time needed: {1:.2f} min".format(k+1,(end-start)/60))
             return True,final_gen
 
     
     p=constants["proc"]
+    survival_rate=0
     if p>1:
         '''Create list of lists of p elements to be used as arguments (population number) in pool.map '''
         list1=[]
@@ -222,25 +175,88 @@ if __name__ == '__main__':
         
         '''Run proc number of populations simultaneously'''    
         a=[]
+        counter=0
         for l in list1:
             if len(l)!=0:
                 pool=Pool(processes=len(l))
-                a.extend(pool.map(main,l))
+                out=pool.map(main,l)
+                a.extend(out)
+                for x,y in enumerate(out):
+                    if y[0]:
+                        survival_rate+=1
+                        f3.write("Population {0} survived!\n".format(x+counter+1))
+                    else:
+                        f3.write("Population {0} died at generation {1}!\n".format(x+counter+1,y[1]))  
+                if survival_rate>=constants["populations"]/2 and constants["stop_half"]:
+                    break
                 pool.terminate()
+                counter+=p
                 
     else:
         a=[]
         for k in range(constants["populations"]):
-             a.append(main(k))
-      
-    survival_rate=0
-    for i,out in enumerate(a):
-        if out[0]:
-            survival_rate+=1
-            f3.write("Population {0} survived!\n".format(i+1))
-        else:
-            f3.write("Population {0} died at generation {1}!\n".format(i+1,out[1]))
-            
-            
-    f3.write("\n\nIn total, {0}/{1} Populations survived.".format(survival_rate,constants["populations"]))
+            out=main(k)
+            a.append(out)
+            if out[0]:
+                survival_rate+=1
+                f3.write("Population {0} survived!\n".format(k+1))
+            else:
+                f3.write("Population {0} died at generation {1}!\n".format(k+1,out[1]))  
+            if survival_rate>=constants["populations"]/2 and constants["stop_half"]:
+                break
+                  
+    f3.write("\n\nIn total, {0}/{1} Populations survived.".format(survival_rate,len(a)))
     f3.close()
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
